requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
defusedxml==0.7.1
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
fonttools==4.58.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
mysqlclient==2.2.7
numpy==2.2.6
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
dj-database-url
gunicorn
